third_party/blink/web_tests/platform/linux/virtual/threaded/external/wpt/css/css-animations/CSSAnimation-canceling.tentative-expected.txt

This is a testharness.js-based test.
PASS Animated style is cleared after canceling a running CSS animation
PASS Animated style is cleared after canceling a filling CSS animation
PASS After canceling an animation, it can still be seeked
PASS After canceling an animation, it can still be re-used
PASS After canceling an animation, updating animation properties doesn't make it live again
PASS After canceling an animation, updating animation-play-state doesn't make it live again
PASS Setting animation-name to 'none' cancels the animation
FAIL Setting display:none on an element cancel its animations assert_equals: expected "idle" but got "running"
FAIL Setting display:none on an ancestor element cancels animations on descendants assert_equals: expected "idle" but got "running"
Harness: the test ran to completion.
